Fix to spawn visual v2. No more preview in the sky.
