+++ Alliance.Client/GameModes/BattleX/BattleGameMode.cs
@@ -1,4 +1,5 @@
-﻿using Alliance.Common.Extensions.FormationEnforcer.Behavior;
+﻿using Alliance.Client.Patch.Behaviors;
+using Alliance.Common.Extensions.FormationEnforcer.Behavior;
 using Alliance.Common.GameModes.Captain.Behaviors;
 using TaleWorlds.Core;
 using TaleWorlds.MountAndBlade;
@@ -19,6 +20,7 @@ namespace Alliance.Client.GameModes.BattleX
                 return new MissionBehavior[]
                 {
                     new FormationBehavior(),
+                    new AllianceAgentVisualSpawnComponent(),
 
                     MissionLobbyComponent.CreateBehavior(),
                     new MultiplayerRoundComponent(),
